pyproject.toml
[project]
name = "grave-border-detection"
version = "0.1.0"
description = "Computer vision project for detecting and vectorizing cemetery grave borders from drone imagery"
readme = "docs/Project_Description.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Berenike Radeck" }]

dependencies = [
    "torch>=2.0",
    "lightning>=2.0",
    "torchvision>=0.15",
    "numpy>=1.24",
    "opencv-python>=4.8",
    "scikit-image>=0.21",
    "albumentations>=1.3",
    "rasterio>=1.3",
    "geopandas>=0.14",
    "shapely>=2.0",
    "pyproj>=3.6",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "tqdm>=4.66",
    "pillow>=10.0",
    "matplotlib>=3.8",
    "segmentation-models-pytorch>=0.5.0",
    "mlflow>=3.8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-mock>=3.12",
    "ruff>=0.1",
    "mypy>=1.7",
    "pre-commit>=3.6",
    "ipykernel>=6.27",
    "jupyter>=1.0",
    "taskipy>=1.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/grave_border_detection"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["grave_border_detection"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "cv2.*",
    "rasterio.*",
    "geopandas.*",
    "shapely.*",
    "albumentations.*",
    "hydra.*",
    "omegaconf.*",
    "skimage.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src/grave_border_detection"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
fail_under = 80

[tool.taskipy.tasks]
lint = "ruff check . && ruff format . && mypy src/"
fix = "ruff check . --fix && ruff format ."
test = "pytest tests/ -v --cov=grave_border_detection"
all = "task lint && task test"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]
